Use jagexappletviewer.jar from oldschool.msi
It seems that this jar is more up to date than the dmg and the jar
download as it has the ability to remove the copyright and language
settings bar after log in.

This makes sense as OSRS is probably the only thing Jagex offers that
actually uses the java appletviewer, whereas RS3 is pretty much all-in
on NXT.

Fixes #127 actually.